requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.13.0
pillow>=9.5.0
numpy>=1.24.0
matplotlib>=3.7.1
scikit-learn>=1.2.2
opencv-python>=4.7.0.72
timm>=0.9.5
gradio>=3.33.0
huggingface_hub>=0.16.4
accelerate>=0.21.0
diffusers>=0.18.0
xformers>=0.0.20
tensorboard>=2.13.0
ipywidgets>=8.0.0
wandb>=0.15.0 